pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "self-repr"
version = "0.1.0"
description = "Mechanistic interpretability: self-representation in LLaMA-3.1-8B-Instruct"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "transformer-lens>=2.0",
    "transformers>=4.40",
    "huggingface-hub>=0.23",
    "numpy>=1.26",
    "scipy>=1.12",
    "scikit-learn>=1.4",
    "h5py>=3.11",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "pandas>=2.2",
    "tqdm>=4.66",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0",
    "ipykernel>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."
